Migrate local download feature to downloadbtn component
